Add scratch card game logic in Day04.kt
Created a new file Day04.kt where a scratch card game is implemented. The file includes two main functions to find the winning numbers on a card and calculate the scores for each round of the game. The functionality is tested with dummy data to ensure it meets the desired criteria.